fix(sm-import): handle eds url domain
