+++ js/import/import.ui.js
@@ -29,9 +29,8 @@ import {
 import * as fragmentUI from '../sections-mapping/sm.ui.js';
 import { buildTransformationRulesFromMapping } from './import.rules.js';
 import TransformFactory from '../shared/transformfactory.js';
-import detectSections from '../sections-mapping/utils.js';
+import detectSections, { getFragmentSectionsMappingData } from '../sections-mapping/utils.js';
 import { preparePagePreview } from '../express/free-mapping/preview-selectors.js';
-import { getFragmentSectionsMappingData } from '../sections-mapping/import/sections-mapping.import.js';
 
 const PARENT_SELECTOR = '.import';
 const CONFIG_PARENT_SELECTOR = `${PARENT_SELECTOR} form`;
@@ -703,13 +702,13 @@ const attachListeners = () => {
               processNext();
             }
             const saveMappingsForAssistant = async () => {
-              let sectionsMapping = getFragmentSectionsMappingData(url);
-              sectionsMapping = JSON.stringify(sectionsMapping, null, 2);
-              if (sectionsMapping) {
+              const sectionsData = getFragmentSectionsMappingData(url);
+
+              if (sectionsData) {
                 if (sessionStorage.getItem(DEMO_TOOL_MODE_SESSION_STORAGE_KEY)) {
-                  await saveBlob(new Blob([sectionsMapping]), 'sections-mapping.json');
+                  await saveBlob(new Blob([JSON.stringify(sectionsData, null, 2)]), 'sections-mapping.json');
                 } else if (dirHandle) {
-                  await saveFile(dirHandle, 'sections-mapping.json', sectionsMapping);
+                  await saveFile(dirHandle, 'sections-mapping.json', JSON.stringify(sectionsData, null, 2));
                 }
               }
             };
